+++ src/features/recovery.sh
@@ -1,6 +1,0 @@
-#!/system/bin/sh
-set -e
-MODDIR=${0%/*}
-. "$MODDIR/../lib/common.sh"
-
-hide_recovery_folders