+++ src/webroot/js/types.ts
@@ -31,7 +31,7 @@ export interface InfoJson {
   security_patch?: string;
   build_patch?: string;
   pif_model?: string;
-  flags?: { twrp?: boolean; blacklist?: boolean; recovery_detected?: boolean };
+  flags?: { twrp?: boolean; blacklist?: boolean };
 }
 
 export interface KeyboxInfoJson {